trading/requirements.txt
alpaca-py
gym-trading-env
